backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
supabase==2.9.1
anthropic>=0.40.0
openai>=1.55.0
slowapi==0.1.9
python-multipart==0.0.12
httpx==0.27.2
tenacity==9.0.0
pdfplumber==0.11.4
python-docx==1.1.2
